🚚 F03: Rename utils folder to ui folders for reusable components
